# WrapAngle optimisation & tests