This is a performance optimisation for the `MathHelper.WrapAngle` method, plus a heap of tests proving that it returns the same results as XNA.

The same result can be calculated using the standard modulo operator which is much faster than the `Math.IEEERemainder` operation. While the 2 operations do give different results (e.g. `10f % MathHelper.TwoPi = 3.716815f` and `Math.IEEERemainder(10f, MathHelper.TwoPi) = -2.566371f`), these are accounted for by the additional limit checks. But don't take my word for it, the tests I've added prove that.

Here are my benchmarks of the original method versus this new one using modulo. The new method takes about 64% less time! But I also added a check to skip the calculations if the value is already within range, and in this case it takes 92% less time than the original!

Method | Mean | StdDev | StdErr | Min | Q1 | Median | Q3 | Max | P95 | Op/s | Scaled | Scaled-StdDev
--- | --- | --- | --- | --- | --- | --- | --- | --- | --- | --- | --- | ---
Original | 51.2411 ns | 0.1013 ns | 0.0281 ns | 51.1090 ns | 51.1818 ns | 51.2110 ns | 51.2752 ns | 51.4786 ns | 51.4391 ns | 19515601.45 | 1.00 | 0.00
New (outside range) | 18.7181 ns | 0.1280 ns | 0.0330 ns | 18.6689 ns | 18.6694 ns | 18.6699 ns | 18.6705 ns | 19.0591 ns | 19.0216 ns | 53424196.85 | 0.36 | 0.00
New (within range) | 4.1722 ns | 0.0004 ns | 0.0001 ns | 4.1715 ns | 4.1721 ns | 4.1722 ns | 4.1724 ns | 4.1729 ns | 4.1727 ns | 239682313.82 | 0.08 | 0.00