@theZMan 2017-01-15
Nice - I think with a slight tweak you can lose one of the branches - not sure if it will give you any more gains though. See http://stackoverflow.com/questions/11498169/dealing-with-angle-wrap-in-c-code