@theZMan 2017-01-15
This one is cool too - no branches at all http://stackoverflow.com/questions/1628386/normalise-orientation-between-0-and-360 not sure if it would give the same results as XNA or not...