@MichaelDePiazzi 2017-01-16
Thanks @theZMan! Unfortunately, neither of these alternative methods gives the same results as XNA.

The first one does seem to work, but it gives slightly different results to XNA mainly due to it having different limits. It's limits are `-Pi <= angle < Pi`, whereas XNA is `-Pi < angle <= Pi`. There's probably also some floating point precision differences due to it modifying the value before getting the remainder. I didn't benchmark it, but I'm guessing the performance difference would be negligible anyway.

The second one I'm assuming you meant the second answer (http://stackoverflow.com/a/2021986)? It definitely seems promising, but unfortunately it has the same issues as the first one (i.e. different limits to XNA). It did inspire me to try this alternative method though (which does give the same results as XNA):
```c#
public static float WrapAngle(float angle)
{
    if ((angle > -Pi) && (angle <= Pi))
        return angle;
    const double divisor = TwoPi;
    const double divisorReciprocal = 1.0 / divisor;
    angle = (float)(angle - (int)(angle * divisorReciprocal) * divisor);
    if (angle <= -Pi)
        return angle + TwoPi;
    if (angle > Pi)
        return angle - TwoPi;
    return angle;
}
```
The performance was basically identical to using modulo though.