@MichaelDePiazzi 2017-01-18
@tomspilman @KonajuGames @dellis1972 - Hey guys, this should be good to merge. I've added pretty extensive unit tests for it, so I'm confident it's working correctly. Thoughts?