@KonajuGames 2017-01-18
I'm happy with this.  Lots of unit tests is good, and better performance too. Merging.